requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
telegraph==1.4.0
APScheduler==3.10.4
lxml==5.1.0
